Add File UTS

+++ com/base/BangunDatar.java
@@ -6,5 +6,5 @@ public interface BangunDatar{
     void tampilluas();
     int panjang();
     int lebar();
-    int luas();
+    double luas();
 }
